cleanup config handling
Now the app buttons and columns are set to appropriate defaults
when no config file is found.

Also the long fields list, used for ordering the json dictionaries,
is removed from the stock config file. The default ordering is
supported by the code now. The config file can still override this
if the order deserves to be changed or more fields are added by
the endpoint programs.